Renamed Actor namespace and set values for Permissions enums
