+++ source/App/source/FunctionApp.SimpleInjector/FunctionApp.SimpleInjector.csproj
@@ -27,7 +27,7 @@ limitations under the License.
 
   <PropertyGroup>
     <PackageId>Energinet.DataHub.Core.App.FunctionApp.SimpleInjector</PackageId>
-    <PackageVersion>7.2.4$(VersionSuffix)</PackageVersion>
+    <PackageVersion>7.2.5$(VersionSuffix)</PackageVersion>
     <Title>Azure Function Dependency Injection Library</Title>
     <Company>Energinet-DataHub</Company>
     <Authors>Energinet-DataHub</Authors>
