+++ source/App/source/FunctionApp/Middleware/ServiceBusActorContextMiddleware.cs
@@ -15,7 +15,7 @@
 using System;
 using System.Diagnostics.CodeAnalysis;
 using System.Threading.Tasks;
-using Energinet.DataHub.Core.App.Common.Abstractions.Actor;
+using Energinet.DataHub.Core.App.Common.Abstractions.Actors;
 using Energinet.DataHub.Core.App.Common.Abstractions.ServiceBus;
 using Energinet.DataHub.Core.App.Common.Parsers.Helpers;
 using Energinet.DataHub.Core.App.FunctionApp.Extensions;
